backend/requirements.txt
fastapi
uvicorn
python-multipart
chromadb
pypdf2
openai
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
sentence-transformers
google-generativeai>=0.3.0
chromadb>=0.4.15
pypdf2>=3.0.1
sentence-transformers>=2.2.2